refactor: game-post-logic
